server/requirements.txt
# Production dependencies for shopping extension server
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
google-generativeai==0.3.2

# Development dependencies (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
